Another round of importing the gradle project into IntelliJ IDEA
The IDE isn't able to handle code completion in the build.gradle file,
despite my setting the Gradle Wrapper distributionUrl to "all"
(to include sources).

On the happier side, the IDE properly picks up libraries from the
build.gradle file so long as I give it a few minutes to sync
and/or to update its plugins.